Moves supports-snapshots driver capability from configuration to code
At the beginning, this is a capability driven by what code can support,
there's no point to set this key in configuration if the driver lacks
the support. As such, this capability should be set by the driver code.

In the future, there should be a field similar to `capabilities`, which
would be used to enable/disable capabilities. If supported by the
driver, it should be still possible to disable a feature.