Adds support for patching of flavor disk info
AWS driver uses `disk.space` property to request coresponding storage
size. This value is set for *custom* flavors, but not for the original,
base flavors. If there is a request to create an instance based on such
flavor, AWS driver would know how much storage space to request.

Note that since this is EBS, it is up to maintainers to pick storage
size for flavors provided by AWS, just like they did for when creating
custom flavors.

```yaml
        - name: t2.small
          cpu:
            family: 6
            family-name: "Haswell"
            model: 63
            model-name: "Intel(R) Xeon(R) CPU E5-2676 v3 @ 2.40GHz"
          disk:
            space: "20 GiB"
```