Adds handling of "no distro tree matches recipe" Beaker error
This error is irrecoverable, and driver should signal that it cannot proceed with provisioning.

Artemis may pick the same pool in the future again as it might be the only viable pool given the request, but that is a different story. By incrementing the metrics and reporting the issue maintainers have a chance to intervene.