Adds caching of image block device mappings
During every provisioning, the driver queried API to fetch image block
device mappings. This information is already available when fetching
pool image info, it is just not cached.

This patch adds caching of block device mappings by simply storing the
value in AWS driver's custom pool image info container, and uses it when
requesting new instance.

That's the gist - the rest is just a bunch of `TypedDict` types to make
sure the JSON blob returned by AWS API is typed properly, and that we're
not trying to store a string where dictionary is expected.