Exposes poolname from tasks where it is known but not a task parameter
There are tasks that operate within the context of a particular pool,
but the pool name is not given to them as a parameter. For example,
`release_pool_data` is given a pool name because there's no guest
request to use as a source - when this task runs, the request is
probably already gone from the DB.

For such tasks, broker middleware can extract the pool name, and expose
it when reporting message durations.

Then there are tasks like `update_guest_request` that do work with a
particular pool, but broker middleware has no access to this
information. For these tasks, we need to expose the pool name, to allow
tracking of these pool-specific actions. The only place where our code
and middleware can exchange information seems to be the message
`options` mapping - therefore, we store pool name in this mapping for
middleware to read it when `poolname` task parameter doesn't exist.