Refactors chain tail handling to be closer to task code
Tail handling - when a chain of tasks runs out of retries - was
implemented with several methods, dispatched from the middleware. It
required a bunch of testing which handling code to use, covering
additional tasks required changing the middleware as well.

This patch refactors the code to use inheritance as much as possible,
and lets us attach tail handler to each affected tasks. Middleware then
just calls the handler without the need to understand it.