Convert guest/snapshot state from string to to enum
The field was storing the enum value, which resulted in `FOO.value` use in
code instead of `FOO` which is more enum-ish. Thinking about where we
need to use `.value` instead of using using enum members everywhere made
the code less readable.

This patch converts DB columns, and places where enum is combined with
the columns.