Converts guest event details column to JSON
A text-like column was used, holding the JSON blob, with our code doing
the conversions (loads/dumps) every time we read/set the data. Both DB
engines we work with can handle JSON natively, there's no reason to
avoid this data type, and it will reduce the area where we could forget
to apply conversion: guest event `details` is supposed to be a JSON
structure, therefore it will always be exactly that, never a string
requiring a `loads` call first.