Adds explicit SSH info to image info structure
At this moment, Artemis expects to use the same username and port when
connecting to all VMs, because all images are same in this regard. This
is not true in general, and we need to teach Artemis that images may be
configured in different ways, which may cover also non-default SSH
username and port.

For this reason, adding a SSH info piece. It is set to default values
now, but in follow up patches, this will change and maintainers will
gain ability to change the values.