Adds support for patch flavor name being a regular expression
Besides the `patch-flavors.[].name`, to patch a single flavor selected
by its full name, this patch adds `patch-flavors.[].name-regex` that,
when used, would patch all flavors with matching names.

This is a handy feature that should shorten amount of "paperwork" needed
to patch flavor families, and it will be used for the future patching of
image infos.