Moves private/public SSH keys into encrypted columns in DB
Instead of an external files, encrypted with Vault, the SSH keys now
live in DB columns, encrypted with the Vault password. The encryption is
handled transparently by SQLAlchemy Utils, not by Ansible Vault.

This move should help use get rid of our dependency on Ansible, the only
purpose was our use of Vault. With encrypted columns, it would be no
longer needed.