test-report: display SSH connection info
When wrapping up the test schedule, print out how to connect to all
those entries.

It's valid only when provisioning was asked not to release guests, and
with more work code could be modified to print the info in such a
situation only, not always - but we would need a guest attribute to
track whether it's supposed to be released or not. This is a simple
low-hanging fruit.