Refactors request to image mapping process to allow "no" as an answer
Sometimes, "no, I don't have suitable image" **is** a valid answer, not
an error. It is not acceptable when acquiring a request, but it's
perfectly fine when asking the pool whether it can handle given request.
Without "no", all negative answers were treated as errors, therefore
pools had only two options: "yes, I can acquire that" or returning a
failure.

And we're starting to see influx of composes that have their own
properties, like ENA support, and can or cannot be used based on the
request - for example, a request for console/URL demands AWS driver to
use an image with ENA support, and what if there is no such image?
Mapping must return "no image" flag, and let driver handle it depending
on the context (`can_acquire()` vs `acquire_guest_request()`).