Adds primitives for constraint pruning and span discovery
Pruning is easy: for a given flavor and constraints, remove constraints
that did not evaluate as true. That should leave only branches that are
satisfied by the given flavor - pick any of them, flavor can deliver.

Spans are harder: generate all (possibly overlapping) combinations of
constraints between the root and leafs of the constraint tree. There is
some magic involved to handle compound constraints correctly. For
example, 2 `Or` nodes under `And` should yield all comibations of children
of the first `Or` with children of the second `Or` - all combinations that
would make `And` true.

When both are combined (prune => spans), on gets all possible
combinations of constraints to satisfy to make the flavor evaluate as
suitable.

For example, following HW requirements:

```yaml
or:
  - memory: >= 4 GiB
  - memory: >= 8 GiB
```

combined with a flavor with 32 GiB of RAM, there are two spans, `memory:
>= 4 GiB` and `memory: >= 8 GiB`. Note that the list of spans is not the
smallest possible - there can be overlap, and the list could be reduced
(`memory: >= 8 GiB` would cover the other branch as well).

More often than not, the list of spans would be short, 1 or 2 items,
like in this example. However, with flavors that can provide additional
disks or NICs, drivers will need to list actual constraints that needs
to be materialized:

```yaml
disk:
  - size: >= 20 GiB
  - size: >= 1 TiB
```

yields span `[disk[0].size >= 10 GiB; disk[1].size >= 1 TiB]`. With
a flavor with fixed root disk of 20 GiB and more, and a possibility to
mount custom volumes of arbitrary size, the second member of this span
tells driver how large this volume needs to be to satisfy the
requirement.