Drops use of Pika parameters in broker connection in favor of URL
Code already accepts URL, but decomposes it to components, because we
want to specify special connection parameters for Pika, like heartbeat.
But, as discovered today, it's possible to specify these parameters with
URL as well, encoded as a query strings.

Therefore droping the special knobs and all related code in favor of
using URL - that way we don't need special knobs for all possible
parameters, all we need is an URL.