Allows drivers to not use the default flavor when there is no suitable one
This is important for `can_acquire()` behavior - when no matching flavor
is found, in the context of `can_acquire()` it's not an error, and
actually it's important to not to pretend the default flavor is a
suitable choice.

Adding a configuration key to control this behavior - it should be
default, but let's not change the default behavior without a feature
switch. If things turn out working nicely, we'd remove the switch.