Adds an `artemis-worker` script to handle running workers everywhere
Dramatiq is fine, but some of the worker aspects are hard to influence.
E.g. queues must be specified as a list of names, prefetch settings can
be changed only via environmetn variables we cannot change, and so on.
On top of that, our tasks package is starting to grow, thanks to tasks
being moved into their own modules for better readability - and updating
the list of modules to import is tedious.

Therefore here's `artemis-worker` that accepts command-line options as
well as environment variables (`--processes` and
`ARTEMIS_WORK_PROCESSES` are checked for input, for example), and takes
care of discovering modules with tasks.