Clears mimatches in names of serialization methods
`serialize_to_json()` was returning Python built-in types which *could*
be easily dumped as a JSON blob, but it was sligthly misleading. On top
of that, there was also `serialize_to_str()` which did dump this "JSON"
into a string.

To make things clear, there's now `serialize()` for conversion to
built-in types, and then `serialize_to_json()` and `serialize_to_yaml()`
for "dump to string as FMT" actions. `serialize_to_str()` is gone, use
one of the available "dumpt to string" methods instead.