Adds customizable template for searching environment/image mappings
When mapping environment to an image, only compose ID is used. This is
good enough in most cases, but there's actually no point in preventing
maintainers to use any other environment attribute.

Patch adds a way how to specify a "needle template" to use instead of
a hardcoded compose ID. The default template is `{{ os.compose }}` which
is backward compatible, but imagine the following template and mapping:

```jinja
{{ os.compose }}:{{ hw.arch }}
```

```yaml
- RHEL-20220411.0:aarch64: RHEL-20220411.0-aarch64
- RHEL-20220411.0:x86_64: RHEL-20220411.0-x86_64
```

Pool would be able to use different images based on the requested architecture.

This could be also simplified when such a split is norm, to avoid one entry per
compose/arch pair in the mapping file:

```yaml
- '(RHEL-\d+\.\d):(aarch64|x86_64)': '\1-\2'
```