Unify infrastructure of server/ and cli/ components
This means mostly Makefile, tox and pre-commit setup.

* Makefiles are now using a shared library for common targets
* tox.ini files have been synced to provide the same level of
  logging and support
* pre-commit is now covering both directories

Note: pre-commit is installed once for each component, it is not shared,
also there are two config files. The final goal is to have just one
covering both server and CLI, that will come later. Running `make -C cli
test-pre-commit` will use CLI's own pre-commit installation and config
file which is a subset of the one in server/ directory. For commits, it
only the server one should be used, even for CLI-only changes.