Fix possible double-free scenario in AWS driver
When releasing resources of a spot instance request, the instance has
been freed first, then the spot instance request. If freeing the
instance has been successfull, but we fail to release the spot request,
whole task would be retried, attempting to free the instance again -
failing, causing code never to get to the spot request ever again.

Using a sequence of tasks for resources that can be released in this
matter, we can free the instance first in the first task, then spot
request in the second task - and if things turn sour for the spot
instance request, only the second task would be retried.