Fixes rendering of Beaker XML translations
When turned into an XML element & rendered, it's possible to loose
information in the process. If `<<` were used as custom variable
delimiters for the template rendering, `<foo bar="<< baz >> phi="pax"/>`
leads to `<foo bar=""/>`, which is obviously wrong. Therefore render
first, then turn the string into a XML object.