Store the last poolname attempted for provisioning a request
This will provide the correct information about what pool we are
switching from in routing. When acquire/update tasks fail to finish
their job, provisioning tail would reset request's poolname - which
means that when routing inspects the request, the previous pool is not
clearly visible. The attribute holds this information.