Azure driver polishing: auth and vm creation/cleanup
With this change authentication via service principal will be supported, which seems to be the right way to do it for artemis driver.

Also guest creation and cleanup has been revised - instead of dumping all azure guests under one resource group and then going through the trouble of meticulously filtering by tags and removing resources one by one a different approach is taken now: a separate resource group is created for each of the guests and the cleanup is as simple as this resource group deletion.