Adds exclusive locking into dispatcher task select query
This should prevent two instances of dispatcher selecting - and
processing - the same table row. `FOR UPDATE` will lock the row - a row
we want to read and remove eventually - and `SKIP LOCKED` will skip any
rows locked by other dispatcher instances. Together, each dispatcher
will select a single row, and it will be row no other dispatcher read
because otherwise it would be locked, and therefore skipped and not
read.