Split "Upgrade to SQLAlchemy 2.0" DB revision into smaller ones
The revision tries to apply several `ALTER TABLE` commands, and by doing
so it risks deadlock with queries started by regular tasks. Which we
observed in our work, and the revision fails to apply or takes too long
for `helm` to finish it.

Splitting the revision into per-table revisions should prevent any
deadlocks - just one table per revision - and let it apply cleanly. The
change should be invisible for existing deployments, it's a bunch of
inserted revisions that seamlessly connect to the "before" and "after"
revisions.