Fixing semantics of "supports kickstart" pool capability
From now on, it's renamed to `supports_native_kickstart`, and when set,
the driver is responsible for delivering kickstart to its backend. When
not set, Artemis will spawn a series of tasks to reinstall the guest
after provisioning.

Drivers that do not support kickstart at all should reject such requests
in their `can_acquire()` method.