Move DB-unrelated code out of transaction in refresh tasks
This is mostly "load pool" or "load pools" step which needs to be
performed in the context of a transaction, but once pool driver
instances are created, DB access is not needed. By wrapping up the
transaction - and marking it read-only - and continuing with whatever
the task needed to do without touching database, we should reduce the
amount of locking and idle-in-transaction connections. This type of
tasks populates the cache, not DB, therefore this should be an easy win.