Collect hostnames from crashed Beaker jobs and avoid them
Similar to "avoid hostnames" and "avoid groups", but instead of
preconfigured entities, these are extracted from aborted and failed
Beaker jobs. Future Beaker jobs for the same guest request should avoid
them.

It is a primitive list, in the future we may "forgive" hostnames in the
way similar to how we forgive unstable pools.