Fix security group leak when instance failed to start
When Artemis creates per-guest security groups, the group is created
first, and then Artemis tries to start an instance. If that fails, the
security group is not reused in the next attempt.

Patch removes this particular leak, by reusing the group in next
attempts from the same pool. There is still a possibility for leak if
the request is routed to another pool, never returning to the former
one, but that should be rare, and it is out of the scope of this PR.