Drivers do not need to set `ProvisioningUpdate.delay_update` to use the default
Drivers had to set the field to a value provided by
`KNOB_UPDATE_GUEST_REQUEST_TICK` for their pool, but since this can be
easily acquired by a task code, it can be moved outside of drivers.
Drivers are still able to set the field to delay the invocation of the
`update-guest-request` task if they wish, but when left unset,
`KNOB_UPDATE_GUEST_REQUEST_TICK` would be used by tasks as the default
value.

Fixes https://issues.redhat.com/browse/TFT-3935.