Introduce `returns` library to replace `gluetool.result`
`returns` is what `gluetool.result` was supposed to become, but that
never happened. `returns` provides not just `Result` container, but
also `Maybe` and `Some`. Together, with all their methods, this package
should help us reduce the amount of `if`s and `return`s we need to write
to handle various results.

Adding the package, and converted one function. For some time, we will
have both `gluetool.result` and `returns.result` in the code, slowly
replacing one with the other.

https://returns.readthedocs.io/en/latest/index.html