More precise control of formatting in `log_dict_yaml`
Formatting larger structures as YAML can be very costly, and it can have
a measurable effect on tasks and their duration. Adding a mechanism to
avoid formatting when the output would not be logged eventually because
of the logging level being too high. This should improve duration of
tasks logging flavors and images, namely `route-guest-request`.